# BUG/PERF: sparse min/max don't densify